embeddings/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic>=2.7
numpy>=1.26
llama-cpp-python>=0.2.90

